added SPHINX/EXAMPLES/etermfig.py changed SPHINX/_templates/layout.html changed SPHINX/celestialbackground.rst changed UTILS/makepack.csh

+++ UTILS/makepack.csh
@@ -1,5 +1,10 @@
 #!/bin/csh
 cd DISTRIBUTION
-\rm -rf build
+if ( ${#argv} >= 1 ) then
+   if (${argv[1]} == "--clean") then
+      echo cleaning build directory
+      \rm -rf build
+   endif
+endif
 python setup.py install --install-lib ../PACKAGE
 cd ..
